feat(e2e): Log number of sent txs (success and failed) (#2328)
When we run e2e's `load` command there is no message telling if the txs
were sent or not, and if not, for what reason (eg. mempool is full, txs
is too large, wrong endpoint, etc.). This PR will log a message every ~1
second with the number of successfully sent and failed transactions.

---

#### PR checklist

- [ ] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec